+++ AsciiTypeGenerator/AsciiService/Controllers/ArtworksController.cs
@@ -2,6 +2,7 @@ using AsciiService.Models.Artwork;
 using AsciiService.Repositories.ArtworksRepository;
 using AsciiService.Shared.Constants;
 using AsciiService.Shared.Constants.Messages;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace AsciiService.Controllers;
@@ -30,6 +31,7 @@ public class ArtworksController(IArtworksRepository artworksRepository)
         return Ok(ArtworkDetailsDto.FromEntity(artwork));
     }
 
+    [Authorize]
     [HttpPost(ApiRoutes.Artworks.Create)]
     public async Task<ActionResult<ArtworkDetailsDto>> CreateArtwork([FromBody] ArtworkUpsertDto request)
     {
@@ -39,15 +41,15 @@ public class ArtworksController(IArtworksRepository artworksRepository)
         if (!ModelState.IsValid)
             return BadRequest(ModelState);
 
-        // TODO: Get author ID from authentication context
-
+        var userId = User.FindFirst("id")?.Value;
         var now = DateTime.UtcNow;
-        var artwork = await artworksRepository.AddAsync(request.ToEntity(null, now, now));
+        var artwork = await artworksRepository.AddAsync(request.ToEntity(userId, now, now));
 
         return CreatedAtAction(nameof(GetArtworkById), new { id = artwork.Id },
             ArtworkDetailsDto.FromEntity(artwork));
     }
 
+    [Authorize]
     [HttpPut(ApiRoutes.Artworks.Update)]
     public async Task<ActionResult<ArtworkDetailsDto>> UpdateArtwork([FromRoute] int id,
         [FromBody] ArtworkUpsertDto request)
@@ -61,24 +63,29 @@ public class ArtworksController(IArtworksRepository artworksRepository)
         if (!await artworksRepository.Exists(id))
             return NotFound(Messages.Error.ArtworkNotFound(id));
 
-        // TODO: Check the author is the same as the one who created the artwork
-
         var artwork = await artworksRepository.GetAsync(id);
+
+        if (artwork.AuthorId != User.FindFirst("id")?.Value)
+            return Forbid(Messages.Error.ForbiddenUpdateArtwork(id));
 
         var artworkUpdate =
             await artworksRepository.UpdateAsync(request.ToEntity(id, artwork.AuthorId, artwork.CreatedAt,
                 DateTime.UtcNow));
 
-        return Ok(artworkUpdate);
+        return Ok(ArtworkDetailsDto.FromEntity(artworkUpdate));
     }
 
+    [Authorize]
     [HttpDelete(ApiRoutes.Artworks.Delete)]
     public async Task<IActionResult> DeleteArtwork([FromRoute] int id)
     {
         if (!await artworksRepository.Exists(id))
             return NotFound(Messages.Error.ArtworkNotFound(id));
 
-        // TODO: Check the author is the same as the one who created the artwork
+        var artwork = await artworksRepository.GetAsync(id);
+
+        if (artwork.AuthorId != User.FindFirst("id")?.Value)
+            return Forbid(Messages.Error.ForbiddenDeleteArtwork(id));
 
         await artworksRepository.DeleteAsync(id);
